Remove unused overloads that do not compile
The `ReadHeader()` overloads taking a character pointer (filename) as a
first parameter instead of a `FILE*` directly do not compile, as they
are calling an overload of `ReadHeader(FILE*)` which does not exist.
Clang does not seem to care as these functions are unused when compiling
MeshLab (and therefore also VCGLib) but GCC tries to compile these
(unused) functions and rightly fails.

This commit simply removes the overloads taking a filename (`char*`).
They are in a header, but since they should fail to compile on *any*
platform I guess it is safe to remove these overloads, as I do not
expect them to be used anywhere successfully.