Edge orientation coherence
Sort edges in output of IntersectionPlaneMesh() to remove duplicate vertex and maintain orientation coherence in the edges.

Before this change, the vertex list contains duplicates and the edge list is not sorted, so there is no easy way to build the polylines of the cut. Calling RemoveDuplicateVertex() will produce a edge list with no coherence, like this example:

```
edge [0 1] goes from [0.12843863 0.38690682 0.1] to [0.13383933 0.3839188  0.1]
edge [2 3] goes from [0.14307424 0.38100217 0.1] to [0.13592989 0.38318165 0.1]
edge [3 1] goes from [0.13592989 0.38318165 0.1] to [0.13383933 0.3839188  0.1]
```

The output is correct, but somehow confusing because edges in the polyline(s) are not in order. After the proposed change, the output will be: 

```
edge [0 1] goes from [0.12843863 0.38690682 0.1] to [0.13383933 0.3839188  0.1]
edge [1 3] goes from [0.13383933 0.3839188  0.1] to [0.13592989 0.38318165 0.1]
edge [3 2] goes from [0.13592989 0.38318165 0.1] to [0.14307424 0.38100217 0.1]
```